finalised refactoring of user_layer into separate modules/packages
