+++ lib9ml/python/nineml/user_layer/random.py
@@ -1,0 +1,10 @@
+from .components import BaseComponent
+
+
+class RandomDistribution(BaseComponent):
+
+    """
+    Component representing a random number distribution, e.g. normal, gamma,
+    binomial.
+    """
+    abstraction_layer_module = 'random'
